fix: Ask for location permissions on home screen after login
